requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pillow>=10.0.0
torch>=2.2.0
torchvision>=0.17.0
